Outsourced getRootEvents and getPublishedEvents
- getRootEvents and getPublishedEvents is used in multiple routes
- Therefore they got outsourced to lib/event/utils.ts
