+++ app/routes/organization/$slug/index.tsx
@@ -17,8 +17,8 @@ import ExternalServiceIcon from "~/components/ExternalService/ExternalServiceIco
 import InputImage from "~/components/FormElements/InputImage/InputImage";
 import { H3 } from "~/components/Heading/Heading";
 import { ExternalService } from "~/components/types";
-import { getEventById } from "~/event.server";
 import { getImageURL } from "~/images.server";
+import { getPublishedAndSortedEvents, getRootEvents } from "~/lib/event/utils";
 import { getOrganizationInitials } from "~/lib/organization/getOrganizationInitials";
 import { getFullName } from "~/lib/profile/getFullName";
 import { getInitials } from "~/lib/profile/getInitials";
@@ -169,47 +169,11 @@ export const loader: LoaderFunction = async (args) => {
       (member) => member.profileId === currentUser.id && member.isPrivileged
     );
   }
-  // Only show root events
   if (organization.responsibleForEvents !== undefined) {
-    let rootEvents: { event: Awaited<ReturnType<typeof getEventById>> }[] = [];
-    await Promise.all(
-      organization.responsibleForEvents.map(async (item) => {
-        let rootItem: { event: Awaited<ReturnType<typeof getEventById>> } = {
-          event: item.event,
-        };
-
-        while (
-          rootItem.event !== null &&
-          rootItem.event.parentEventId !== null
-        ) {
-          rootItem.event = await getEventById(rootItem.event.parentEventId);
-        }
-        if (
-          !rootEvents.some((item) => {
-            if (item.event !== null && rootItem.event !== null) {
-              return item.event.slug === rootItem.event.slug;
-            } else {
-              return false;
-            }
-          })
-        ) {
-          rootEvents.push(rootItem);
-        }
-      })
-    );
+    // Only show root events
+    const rootEvents = await getRootEvents(organization.responsibleForEvents);
     // Only show published events
-    //@ts-ignore
-    organization.responsibleForEvents = rootEvents
-      .filter((item) => {
-        return item.event !== null && item.event.published;
-      })
-      .sort((a, b) => {
-        if (a.event !== null && b.event !== null) {
-          return a.event.name.localeCompare(b.event.name);
-        } else {
-          return 0;
-        }
-      });
+    organization.responsibleForEvents = getPublishedAndSortedEvents(rootEvents);
   }
 
   return {
